[ne2k] Issue mfld-fr#3: ETH driver skeleton
